+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## [4.3.X.X] - XXXX-XX-XX
 ## Added
 - Connectivity status of depots
+- OTP field to login dialog
+- OTP option to command line
 
 ## Fixed
 - SWAudit command line options